+++ jvm/src/main/scala/com/nawforce/apexlink/cst/Primaries.scala
@@ -19,11 +19,7 @@ import com.nawforce.apexlink.org.Referenceable
 import com.nawforce.apexlink.types.apex.{ApexClassDeclaration, ApexFieldLike}
 import com.nawforce.apexlink.types.core.{FieldDeclaration, TypeDeclaration}
 import com.nawforce.apexlink.types.platform.PlatformTypes
-import com.nawforce.pkgforce.modifiers.{
-  INTEGRATION_TEST_ANNOTATION,
-  PRIVATE_MODIFIER,
-  TEST_VISIBLE_ANNOTATION
-}
+import com.nawforce.pkgforce.modifiers.PRIVATE_MODIFIER
 import com.nawforce.pkgforce.names._
 import com.nawforce.pkgforce.path.Location
 import com.nawforce.runtime.parsers.CodeParser
@@ -155,13 +151,14 @@ final case class IdPrimary(id: Id) extends Primary {
 
     if (field.nonEmpty && isAccessible(td, field.get, staticContext)) {
       if (
-        field.get.modifiers.contains(TEST_VISIBLE_ANNOTATION) &&
+        field.get.isTestVisible &&
         field.get.visibility.contains(PRIVATE_MODIFIER) &&
-        context.thisType.modifiers.contains(INTEGRATION_TEST_ANNOTATION)
+        !field.get.thisTypeIdOpt.contains(context.typeId) &&
+        !context.thisType.isUnitTest
       ) {
         context.logError(
           location,
-          "@IntegrationTest classes can not access private @TestVisible fields"
+          "Private @TestVisible fields can only be accessed from @IsTest classes"
         )
       }
       Referenceable.addReferencingLocation(field.get, location, context.thisType)