+++ jvm/src/test/scala/com/nawforce/apexlink/cst/MethodTest.scala
@@ -125,8 +125,32 @@ class MethodTest extends AnyFunSuite with TestHelper {
     assert(
       getMessages(
         root.join("Dummy.cls")
-      ) == "Error: line 1 at 79-94: @IntegrationTest classes can not access private @TestVisible methods\n"
+      ) == "Error: line 1 at 79-94: Private @TestVisible methods can only be accessed from @IsTest classes\n"
     )
+  }
+
+  test("Non-test class can not access private TestVisible method") {
+    typeDeclarations(
+      Map(
+        "Target.cls" -> "public class Target {@TestVisible private static void helper(){}}",
+        "Dummy.cls"  -> "public class Dummy {public static void f2() {Target.helper();}}"
+      )
+    )
+    assert(
+      getMessages(
+        root.join("Dummy.cls")
+      ) == "Error: line 1 at 45-60: Private @TestVisible methods can only be accessed from @IsTest classes\n"
+    )
+  }
+
+  test("IsTest class can access private TestVisible method") {
+    typeDeclarations(
+      Map(
+        "Target.cls" -> "public class Target {@TestVisible private static void helper(){}}",
+        "Dummy.cls" -> "@IsTest public class Dummy {@IsTest public static void f2() {Target.helper();}}"
+      )
+    )
+    assert(getMessages(root.join("Dummy.cls")).isEmpty)
   }
 
   test("IntegrationTest can not access private TestVisible field") {
@@ -139,8 +163,32 @@ class MethodTest extends AnyFunSuite with TestHelper {
     assert(
       getMessages(
         root.join("Dummy.cls")
-      ) == "Error: line 1 at 94-107: @IntegrationTest classes can not access private @TestVisible fields\n"
+      ) == "Error: line 1 at 94-107: Private @TestVisible fields can only be accessed from @IsTest classes\n"
     )
+  }
+
+  test("Non-test class can not access private TestVisible field") {
+    typeDeclarations(
+      Map(
+        "Target.cls" -> "public class Target {@TestVisible private static String helper;}",
+        "Dummy.cls" -> "public class Dummy {public static void f2() {String value = Target.helper;}}"
+      )
+    )
+    assert(
+      getMessages(
+        root.join("Dummy.cls")
+      ) == "Error: line 1 at 60-73: Private @TestVisible fields can only be accessed from @IsTest classes\n"
+    )
+  }
+
+  test("IsTest class can access private TestVisible field") {
+    typeDeclarations(
+      Map(
+        "Target.cls" -> "public class Target {@TestVisible private static String helper;}",
+        "Dummy.cls" -> "@IsTest public class Dummy {@IsTest public static void f2() {String value = Target.helper;}}"
+      )
+    )
+    assert(getMessages(root.join("Dummy.cls")).isEmpty)
   }
 
   test("Method call with non-ambiguous target") {